% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vin_format.R
\name{check_vin_format}
\alias{check_vin_format}
\title{Check VIN Length and Characters}
\usage{
check_vin_format(vin)
}
\arguments{
\item{vin}{A character. Should be a properly formatted Vehicle Identification Number.
Wildcards (e.g., '*') are acceptable.}
}
\value{
Logical. TRUE if the VIN is 17 chars and does not contain \verb{[IOQ]}, FALSE otherwise.
}
\description{
Checks that VINs are 17 characters long and do not include any prohibited characters
(I, O, Q).
}
\examples{
# Random VIN
check_vin_format("3VWLL7AJ9BM053541")
# With wild card
check_vin_format("3VWLL7AJ9BM*53541")
}
