% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vin.R
\name{check_vin_purrr}
\alias{check_vin_purrr}
\title{Verify VIN Validity Using Purrr}
\usage{
check_vin_purrr(vin, guess = FALSE)
}
\arguments{
\item{vin}{A character vector of VINs to check. Wildcards (e.g. *) are NOT allowed.}

\item{guess}{Logical. Should values for illegal characters be guessed?}
}
\description{
Verify VIN Validity Using Purrr
}
