% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{swap_letter}
\alias{swap_letter}
\title{Replace a Letter in a Character Vector}
\usage{
swap_letter(.string, .target, .replacement)
}
\arguments{
\item{.string}{character vector}

\item{.target}{character to replace}

\item{.replacement}{character to substitute}
}
\value{
the modified string
}
\description{
Replace a Letter in a Character Vector
}
