% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{swap_map}
\alias{swap_map}
\title{Replace Multiple Letters in a Character Vector}
\usage{
swap_map(.string, .targets, .replacements)
}
\arguments{
\item{.string}{character vector}

\item{.targets}{characters to replace}

\item{.replacements}{characters to substitute}
}
\value{
the modified string
}
\description{
Replace Multiple Letters in a Character Vector
}
