% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_digit_valid.R
\name{check_digit_valid}
\alias{check_digit_valid}
\title{Check for Valid VIN Check Digit}
\usage{
check_digit_valid(vin, value = FALSE)
}
\arguments{
\item{vin}{character. The VIN to check. VINs must be compete, i.e. 17 digits
with no wildcards.}

\item{value}{logical. Should the calculated check digit be returned instead?}
}
\value{
logical. The result of comparing the calculated check digit with the VIN position 9
}
\description{
Calculates the VIN check digit and compares it to VIN position 9.
For US-based VINs, this determines if the VIN is valid.
This may not apply to VINs from outside of the United States.
}
\examples{
check_digit_valid("WDBEA30D3HA391172") # True
check_digit_valid
}
